requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
winotify
pyinstaller
pystray
Pillow
requests
pyperclip
psutil
